more accurate webhooks seen
i was tracking webhooks even if they didn't see, this will provide more accurate tracking.

this is completely useless in terms of proxy functionality and is mainly just for the funny numbers on the website. requests will stay the same though as regardless of the result it is still a request processed by the proxy!